artadoors/requirements.txt
asgiref==3.8.1
Django==5.1.2
django-formtools==2.5.1
django-js-asset==2.2.0
django-mptt==0.16.0
django-mptt-admin==2.7.0
pillow==11.0.0
pysqlite3-binary==0.5.4
sqlparse==0.5.1
